Safari-611.1.10/LayoutTests/fast/dom/timer-height-change-no-throttling-expected.txt

Tests that a repeating timer changing the height of an element outside viewport does not get throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isTimerThrottled(timeoutId) is false
PASS internals.isTimerThrottled(timeoutId) is false
PASS successfullyParsed is true

TEST COMPLETE
TEST
